Remove Win32 ANSI calls from DiskInteface
Ensure Win32 Unicode APIs are called when trying to
perform real disk i/o. This is necessary to properly
support long file paths on Windows.

For example, FindFirstFileExW() used in RealDiskInterface
has not MAC_PATH restriction, though FindFirstFileExA()
still does, even when long path support is enabled
on the host machine.

Note that generally speaking, it is unknown whether Windows
CRT functions, such as fopen(), rename(), _unlink(), always
properly support long file paths (it may very well depend
on the version of MSVCRT linked to the executable), so
it is better to err on the side of caution and always
try to use the wide-char versions of these functions,
when available, or to fall back to Unicode Win32 API
functions otherwise.

This applies here and the commits following this one
in the same pull request.